hotel/requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-bootstrap-datepicker-widget==0.1
idna==2.8
oauthlib==3.0.2
pkg-resources==0.0.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
urllib3==1.25.3
